Move `LocalServer` to `local-jdbc` module
